Hickle v2 now passes most unit tests
Still a few unit tests to go, and will need
a tidy-up afterwards, but mostly complete.

TODO: tidy up new code
TODO: Add legacy support for old hickle files
TODO: get final unit tests passing
